+++ lib/EasyRdf.php
@@ -105,6 +105,11 @@ require_once "EasyRdf/Literal/DateTime.php";
 require_once "EasyRdf/Literal/Decimal.php";
 
 /**
+ * @see EasyRdf_Literal_HexBinary
+ */
+require_once "EasyRdf/Literal/HexBinary.php";
+
+/**
  * @see EasyRdf_Literal_Integer
  */
 require_once "EasyRdf/Literal/Integer.php";