+++ lib/EasyRdf/Parser/Json.php
@@ -5,7 +5,7 @@
  *
  * LICENSE
  *
- * Copyright (c) 2009-2010 Nicholas J Humfrey.  All rights reserved.
+ * Copyright (c) 2009-2012 Nicholas J Humfrey.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -31,7 +31,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @package    EasyRdf
- * @copyright  Copyright (c) 2009-2010 Nicholas J Humfrey
+ * @copyright  Copyright (c) 2009-2012 Nicholas J Humfrey
  * @license    http://www.opensource.org/licenses/bsd-license.php
  * @version    $Id$
  */
@@ -42,11 +42,53 @@
  * http://n2.talis.com/wiki/RDF_JSON_Specification
  *
  * @package    EasyRdf
- * @copyright  Copyright (c) 2009-2010 Nicholas J Humfrey
+ * @copyright  Copyright (c) 2009-2012 Nicholas J Humfrey
  * @license    http://www.opensource.org/licenses/bsd-license.php
  */
 class EasyRdf_Parser_Json extends EasyRdf_Parser_RdfPhp
 {
+    private $_jsonLastErrorExists = false;
+
+    /**
+     * Constructor
+     *
+     * @return object EasyRdf_Parser_Json
+     */
+    public function __construct()
+    {
+        $this->_jsonLastErrorExists = function_exists('json_last_error');
+    }
+
+    /** Return the last JSON parser error as a string
+     *
+     * If json_last_error() is not available a generic message will be returned.
+     *
+     * @ignore
+     */
+    protected function _jsonLastErrorString()
+    {
+        if ($this->_jsonLastErrorExists) {
+            switch (json_last_error()) {
+                case JSON_ERROR_NONE:
+                   return null;
+                case JSON_ERROR_DEPTH:
+                   return "JSON Parse error: the maximum stack depth has been exceeded";
+                case JSON_ERROR_STATE_MISMATCH:
+                   return "JSON Parse error: invalid or malformed JSON";
+                case JSON_ERROR_CTRL_CHAR:
+                   return "JSON Parse error: control character error, possibly incorrectly encoded";
+                case JSON_ERROR_SYNTAX:
+                   return "JSON Parse syntax error";
+                case JSON_ERROR_UTF8:
+                   return "JSON Parse error: malformed UTF-8 characters, possibly incorrectly encoded";
+                default:
+                   return "JSON Parse error: unknown";
+            }
+        } else {
+           return "JSON Parse error";
+        }
+    }
+
     /** Parse the triple-centric JSON format, as output by libraptor
      *
      * http://librdf.org/raptor/api/serializer-json.html
@@ -102,10 +144,9 @@ class EasyRdf_Parser_Json extends EasyRdf_Parser_RdfPhp
         $this->resetBnodeMap();
 
         $decoded = @json_decode(strval($data), true);
-        if (!$decoded) {
-            # FIXME: can we get an error message for json_decode?
+        if ($decoded === null) {
             throw new EasyRdf_Exception(
-                "Failed to parse RDF/JSON"
+                $this->_jsonLastErrorString()
             );
         }